Use a systemd unit to manually set qdisc for eth0.
The previous approach didn't work as the .config file needs to be
written before sysctl parameters are read, and that's not the case.

I could have added a sysctl --system just after that, but at that
point we'd still need to reconfigure eth0 by running tc (as in the
current commit), which in the meantime took the wrong configuration.
I guess just running tc to replace the default during the boot
process works as intended and it's an overall smaller change.