Causes shim.sh to exit with exit code of plugin
Previously, shim.sh would always exit with status code 0. In this
scenario, if the plugin that shim.sh called exited with a non-zero
status, indicating failure, shim.sh would still exit with 0, causing the
caller (the container runtime) to think everything succeeded, when it
really didn't. This commit simply causes shim.sh to explicitly exit
with the same exit code of the called plugin.