Adds new max-rate.service
This new service will determine the ndt-server max-rate and write the
result to a file. This file will be used by the setup_k8s.sh script
which runs after a machine boots and will set a pod label which can be
used by the ndt-server to correctly set the -txcontroller.max-rate flag.