Reverts file to original state, plus one error handler
I mistakenly confused the MAXRATE variable in this file with the
ndt-server -txcontroller.max-rate flag, when in fact they are not
related in anyway way. This commit restores this file to the way it was
before, but adds small bit of error checking so that the unit will exit
with a failure if the `tc` commands fails for some reason.