Changes directory for ndt max-rate value file
The previous file of /etc/ndt-max-rate does not coincide with a location
used for anything else on the platform. Since we already have a
precendent of putting some data in /var/local/metadata, then just use
that directory. Additionally the file was renamed from "ndt-max-rate" to
"iface-max-rate". ndt-server can use the value, but so could any other
service, so now the name is more generic.