Adds new systemd service to write NDT max-rate to a file (#219)
* Adds new max-rate.service

This new service will determine the ndt-server max-rate and write the
result to a file. This file will be used by the setup_k8s.sh script
which runs after a machine boots and will set a pod label which can be
used by the ndt-server to correctly set the -txcontroller.max-rate flag.

* Adds an [Install] section so unit get started

* Sets max-rate.service to run after network-online.target

* Starts max-rate.service after networkd is up

* Sets configure_tc_fq.sh to use local max-rate value

Previously, this script did some back bending to figure out what the NDT
max-rate should be. Now it relies on a new service named
max-rate.service to drop the max rate into a local file
(/etc/ndt-max-rate).

* Sets max-rate.service to run after nss-lookup.target

* Adds back the [Install] section to max-rate service

* Adds a loop looking for name resolution

* Adds SITEINFO_URL var, removes grep from nslookup

* Loops till DNS resolution is functional

* Redirects nslookup output to /dev/null.. less noisy

* Fixes variable name, adds error checking

* Reverts file to original state, plus one error handler

I mistakenly confused the MAXRATE variable in this file with the
ndt-server -txcontroller.max-rate flag, when in fact they are not
related in anyway way. This commit restores this file to the way it was
before, but adds small bit of error checking so that the unit will exit
with a failure if the `tc` commands fails for some reason.

* Changes directory for ndt max-rate value file

The previous file of /etc/ndt-max-rate does not coincide with a location
used for anything else on the platform. Since we already have a
precendent of putting some data in /var/local/metadata, then just use
that directory. Additionally the file was renamed from "ndt-max-rate" to
"iface-max-rate". ndt-server can use the value, but so could any other
service, so now the name is more generic.

* Removes dependency on configure-tc-fq.service

This service does not use the ndt-server max-rate for its configuration,
so can run whenever.

* Adds docstring + loosens loop iteration to 0.1s

The docstring briefly describes what the script does and why it exists.
The previous loop was as tight as the system would allow. This commit
puts a `sleep .1` inside the loop so that it doesn't needlessly run so
fast.