Updates k8s to v1.21.14 + exports new metric about fq qdisc config status (#222)
* Updates k8s to v1.21.14 + upgrade CNI,flannel,crictl

* Modifies configure_tc_fq script to write out metric

This is a small modification to the configure_tc_fq.sh script, which
causes it to write out a node_exporter textfile collector metric file so
that we can monitor whether the qdisc is setup correctly on each machine
at boot time.

* Moves temp metric file after any fatal error

Also, converts tc "maxrate" values to a bytes-per-second value, since
this is how the value is stored internally, and reported by tc when
queried.

* Adds a few bug fixes and improvements to qdisc script

* Configures qdisc setup to run after cache-data.mount

configure-tc-fq.service writes a file to /cache/data, so make sure that
the mount exists.

* Makes sure /cache/data/node-exporter exists

* Renames function and updates comment

* Updates calls to renamed function