Adds back MLXROM_VERSION + special MLX make cflags
It seems that the MFT "flint" has special requirements for the version
string, so this commit reverts the string to the way it was before, even
though it no longer has any meaning (we are not using flexboot), but
ipxe sources to build ROM images. It also seems that "flint" requires
some special string in the image as a "signature" that it validates.

This commit basically falls back on doing the same thing the
stage3_mlxrom build did using the flexboot sources, but instead does the
same with the ipxe sources.