Builds stage1 mrom images using iPXE sources (#228)
* Uses iPXE source code instead of flexboot

The "flexboot" code fro Mellanox is not maintained for ConnectX-3 cards,
and the most recent version is 6 years old. iPXE has native support for
the ConnectX-3 card. This commit makes changes to the stage1_mlxrom
builds to use the iPXE source code instead.

* Updates MFT to latest version 4.22.0-96

MFT is Mellanox Firmware Tools.

* Adds --yes flag to all flint and mlxconfig commands

* Adds back MLXROM_VERSION + special MLX make cflags

It seems that the MFT "flint" has special requirements for the version
string, so this commit reverts the string to the way it was before, even
though it no longer has any meaning (we are not using flexboot), but
ipxe sources to build ROM images. It also seems that "flint" requires
some special string in the image as a "signature" that it validates.

This commit basically falls back on doing the same thing the
stage3_mlxrom build did using the flexboot sources, but instead does the
same with the ipxe sources.

* Patches config/general.h to configure build options

* Fixes config.h patch so it can apply cleanly

* Adds ipxe source as a git submodule

* Updates submodules, instead of cloning repos

* Adds --init flag to git-submodule command