Increases build timeout for stage1 images to 10h
For a long time the build timeout was 4h. Looking back at the build
history, stage1 build times were hovering around 3h30m, give or take.
In the past few months stage1 builds started timing out at 4h, but we
didn't notice because we weren't receiving build failure alerts to the
oti-alerts Slack channel. But now we do get alerts there and noticed. I
raised the timeout to 5h, but even that timed out on the most recent
production build. I don't know why things are taking so much longer, but
this commit raises the timeout to 10h, which should be plenty by 4 to 5
hours.