Sets proper ownership of mount-data subdirs
Each subdirectory of then /mnt/local persistent disk on internal cluster
VMs will have unique ownership, depending on what user the process is
running as.

Additionally, for internal machines, don't loop on whether the token
server is available, but instead on whether the control plane is ready.
The token-server runs a Docker container and will almost certainly be up
and running before the control plane is initialized, in the cases where
the control plane is being created. This makes sure that the CA cert
hash fetched from poject metadata is the most current one.