Removes application of workloads to init cluster
Applying M-Lab workloads to the newly initialized cluster seems out of
scope for the general task of initializing the control plane, especially
if it requires modified versions of the "create" and "apply" scripts
that exists in the k8s-support repo.

Also, renames the mount-data service and script to mount-data-internal
so that the names align with the join-cluster-internal script and
service, and so the name better implies what they are for.