Uploads k8s admin.conf to GCS
This should be considered a bug, but it is what we have been doing for
years now. This is a stop-gap until we find a better way for Cloud Build
to authenticate with the cluster to apply workloads during k8s-support
builds.

Also, apply the flannel-physical DaemonSet on cluster init, in addition
to flannel-virtual.

Also, use jq's long flag --raw-output over the short flag -r.