Configures node name to be more like DNS
This should allow us to strip the last 5 characters of the node name to
get the name of the load balancer for the MIG. This could be useful for
any number of things. For example, the uuid-annotator can use the node
name (minus the last 5 characters) to find the name of the load balancer
in siteinfo's annotations.json so it knows how to annotate data on this
VM instance.

Fixes a bug where $lb_dns was being referenced before being declared.