Removes writing hostname to a metadata file
This functionality was added thinking that uuid-annotator, and possibly
other services, would use the value in this file to know the public name
of the load balancer (behind which the VM sits). But instead, this was
handled by updating the m-lab/go/host package to better support MIG
instance names, and then things like heartbeat, pusher, uuid-annotator,
etc. were updated to rely on this updated package to get the appropriate
host name.