Automatically create the control plane, when it doesn't already exist (#234)
* Adds two new services to create control plane

The two new services, mount-cluster-data.service and
create-control-plane.service, will mount an attached disk to a known
location, and will initialize the k8s cluster, if necessary.

* Fixes the scripts that create the k8s cluster

* Adds new func to add a machine to the load balancer

Machines should not be added to the load balancer until the API
components are up and running. See the comments in the code changes for
this commit for reasoning.

Additionally, a few small tweaks to the rootfs for convenience.

* Various fixes to create-control-plane service

* Fixes for create-control-plane service and script

* Adds new "internal" custom image

This new "internal" image will be used on platform cluster machines that
run inside of GCP, are not control plane nodes, but provide other
critical funcationality, like hosting Prometheus, etc. There is very
little difference between it and the usual platform cluster image other
than it joins the cluster in a different way, using internal GCP
resources and services.

* Try to get a bootstrap token forever

This commit modifies the join-cluster and join-cluster-internal scripts
to try to get a bootstrap token indefinitely. There is no reason to stop
trying, and failure to get a token could be due to something transitory,
such as the control plane rebooting or initializing, etc.

* Adds proper .sh suffix to cluster-init shell scripts

* Don't fail script when curl fails

The script should continue even if a curl request fails, so "|| true"
after the curl command.

* Label initial node with mlab/type=vritual early

Without this label, flannel will not be scheduled on the node, causing
other workloads to fail to functions, specifically cert-manager.

* Uses hostname instead of FQDN for Prom machine

When using the FQDN kubeadm was issuing a warning that the name part of
the domain must be less than 64 characters. Somehow this was causing
problems with the local kubelet, and causing it to fail to join the
cluster. This replaces FQDN with hostname.

* Adds a new mount service/script for internal services

Some services, e.g., prometheus, need a persistent disk for state data.
This commit adds a new systemd servide and corresponding script to mount
a persistent volume on the machine at /mnt/local/, and creates a subdir
of that dir named after the service that will use it.

* Renames two packer configuration scripts

The new names are more consistent and easier to read, I think. This
commit also contains a few small fixes.

* Sets proper ownership of mount-data subdirs

Each subdirectory of then /mnt/local persistent disk on internal cluster
VMs will have unique ownership, depending on what user the process is
running as.

Additionally, for internal machines, don't loop on whether the token
server is available, but instead on whether the control plane is ready.
The token-server runs a Docker container and will almost certainly be up
and running before the control plane is initialized, in the cases where
the control plane is being created. This makes sure that the CA cert
hash fetched from poject metadata is the most current one.

* Adds sleep to join-cluster-internal

Makes sure that the control plane initialization is complete.

* Fixes unbound variable + sets KUBECONFIG

* Introduces condition: don't join cluster if already joined

This commit introduces a new condition to the join-cluster and
join-cluster-internal systemd services. If the file
/etc/kubernetes/kubelet.conf exists, then the machine is likely already
joined to the cluster, and these units do not need to be run.

* Fixes unit description for create-control-plane

* Stop using sandbox-kinkade in k8s-support for init

Don't use the sandbox-kinkade branch of k8s-support for apply workloads
to the newly initialized cluster. Hopefully this PR will be merged,
adding the necessary files to the main branch:
https://github.com/m-lab/k8s-support/pull/761

* Removes application of workloads to init cluster

Applying M-Lab workloads to the newly initialized cluster seems out of
scope for the general task of initializing the control plane, especially
if it requires modified versions of the "create" and "apply" scripts
that exists in the k8s-support repo.

Also, renames the mount-data service and script to mount-data-internal
so that the names align with the join-cluster-internal script and
service, and so the name better implies what they are for.

* Fixes name of services in configure_image_interal.sh

* Apply flannel to cluster on initialization

If flannel is not running in the cluster, then cluster networking is not
up, and nodes will not be flagged as Ready.

Also, adds the setup_k8s.sh template to the filesystem, and it will
be evaluated and uploaded to GCS when the cluster is initialized.

* Fixes syntax error in create-control-plane script

Also, breaks Packer builds out from stage3 builds to speed things up.
Packer should be fairly quick, while stage3 builds take nearly 30m
today.

* Look for cert_key in cluster_data metadata field

Also, fixes small sytax error in Packer Cloudbuild config.

* Adds missing steps: field to packer cloudbuild

* Adds ARTIFACTS to build env

* Adds IMAGES_VERSION back to Packer cloudbuild

* Removes sha256 prefix from CA hash string in template

* Uploads k8s admin.conf to GCS

This should be considered a bug, but it is what we have been doing for
years now. This is a stop-gap until we find a better way for Cloud Build
to authenticate with the cluster to apply workloads during k8s-support
builds.

Also, apply the flannel-physical DaemonSet on cluster init, in addition
to flannel-virtual.

Also, use jq's long flag --raw-output over the short flag -r.

* Fixes unbound variable, and renames to lb_dns

Also, get cluster CA cert hash from project metadata rather than trying
to extract it from setup_k8s.sh.

* Requires mount-data-api for token-server.service

Also makes sure it runs after mount-data-api.service, since that unit
creates directories that token-server mounts.

* Checks that $dev_name exist, else fails unit

If the script cannot successfully extract $dev_name from the
cluster_data metadata, then fail the mount-data-api.sh script (and
service), since we require that name for the service to work as
intended.

* Reconfigures join-cluster.sh script as GCP-only

This is only temporary, until we can update token-server to authenticate
requests from outside of GCP.

Additionally, reconfigures the script to work better generally.

* Configures node name to be more like DNS

This should allow us to strip the last 5 characters of the node name to
get the name of the load balancer for the MIG. This could be useful for
any number of things. For example, the uuid-annotator can use the node
name (minus the last 5 characters) to find the name of the load balancer
in siteinfo's annotations.json so it knows how to annotate data on this
VM instance.

Fixes a bug where $lb_dns was being referenced before being declared.

* Writes node name to custom metadata

The value of this file will be used by uuid-annotator to determine which
hostname from siteinfo to use for annotating experiment data.

* Fixes bugs in the write-metadata systemd units

For the virtual one, there was an unbound variable. For the physical
one, it was evaluating the hostname too soon in the boot.

* Removes writing hostname to a metadata file

This functionality was added thinking that uuid-annotator, and possibly
other services, would use the value in this file to know the public name
of the load balancer (behind which the VM sits). But instead, this was
handled by updating the m-lab/go/host package to better support MIG
instance names, and then things like heartbeat, pusher, uuid-annotator,
etc. were updated to rely on this updated package to get the appropriate
host name.