Don't mkdir /etc/kubernetes/manifests on API machines
On control plane machines, /etc/kubernetes is supposed to be a symlink
to /mnt/cluster-data/kubernetes. When /etc/kubernetes already exists as
a regular dir, then ln creates a symlink inside /etc/kubernetes,
breaking the configuration and breakage of the create-control-plane
service. Anyway, on control plane nodes that directory will be created
automatically by kubeadm.