Refactors the join-cluster.sh script
Previously, the script assumed that all VMs were going to be part of a
MIG. We have decided to have a hybrid approach with both MIGs and
standard VMs, which required a few changes.

Additionally, configure the script to the V2 allocate_k8s_token ePoxy
extension, which returns all the data needed to join the cluster, not
just the token. This also required some refactoring of the code.