Use allocate_k8s_token v2 to join cluster (#237)
* Uses CA cert hash from ePoxy extension server

Version 2 of the allocate_k8s_token returns all data necessary to join
the cluster, so there is no need to upload the CA cert hash to project
metadata.

A few other improvements and fixes are added to this:
* Run create-control-plane.service on every boot, which allows the
  script to add the control plane instances to the appropriate instance
  group, if necessary.
* Renames lb_dns and token_server_address to api_load_balancer and
  epoxy_extension_server, respectively.

* Fixes typo in shell variable name

* Fixes bash unbound variable

* Restores --container-runtime-endpoint flag to kubelet

* Moves all code inside of main()

It should be there anyway, and the script was attempting to run a
function before it was defined in the file.

* Mount PD before starting kubelet

If the persistent disk is not mounted, then
/var/lib/kubelet/kubeadm-flags.env is not available, which containes the
flag --container-runtime-endpoint pointing to containerd. Since control
plane nodes also run dockerd, the kubelet apparently defaults to using
docker as the CRI instead of containerd, and then things don't work.