Changes systemd DefaultTimeoutStopSec to 30s
Now that we are using the heartbeat service to inform Locate of a
service's status, we no longer need the old 180s setting. That old,
high value was only there to compensate for the delay of getting status
to mlab-ns. The new 30s value is only to allow running NDT tests to
finish gracefully.