Adds a new "leave-cluster" service and script (#241)
* Adds new leave-cluster.service and script

This will allow VMs to leave the cluster before they get shutdown. This
is important for members of MIGs, which may get scaled out and back in,
which would otherwise leave an orphaned node in the cluster in a
NotReady state.

* Changes systemd DefaultTimeoutStopSec to 30s

Now that we are using the heartbeat service to inform Locate of a
service's status, we no longer need the old 180s setting. That old,
high value was only there to compensate for the delay of getting status
to mlab-ns. The new 30s value is only to allow running NDT tests to
finish gracefully.

* Fixes an unbound variable in leave-cluster script

* Sets KUBECONFIG env variable for extension-server

allocate_k8s_token was able to work without this because kubeadm will
look in some well known locations for a valid kubeconfig and would find
the one in /etc/kubernetes. Not so for the kubelet in the "delete"
extension.

* Make leave-cluster.sh wait for the API to be up

The epoxy-extension-server runs on the API nodes, and if the API isn't
up, then likely neither is the epoxy-extension-server.

Also, order the systemd unit _after_ network.target, which should ensure
on shutdown this script will get run before systemd takes down
networking.

* Fixes an unbound variable in leave-cluste.sh

* Small changes per PR feedback

* Improves/fixes language in a few comments.
* Passes the control plane CA certificate to curl so that we don't have
  to use the `--insecure` flag.