Uses list instead of describe for gcloud operation
When using "describe" gcloud will exit with an error if the machine
doesn't exist, causing the script to exit. This commit simply changes
the "describe" operation to a "list" operation with a filter. In this
latter case, even if the filter doesn't find anything, gcloud exits
gracefully and the script can continue the loop waiting for the VM to
exist.