Adds cert_key to project metadata
Previously, the cert_key was added directly to the "cluster_data"
metadata entry for non-init control plane machines. However, this
required the init control plane machine script to loop, waiting for
those other machines to be created. Additionally, the process of adding
the cert_key to an existing metadata field was rather ungainly and ugly.
I consider adding the cert_key to project metadata to be a relatively
safe operation. See my comment in create-control-plane.sh for rationale.