Adds cert_key to project metadata  (#243)
* Uses list instead of describe for gcloud operation

When using "describe" gcloud will exit with an error if the machine
doesn't exist, causing the script to exit. This commit simply changes
the "describe" operation to a "list" operation with a filter. In this
latter case, even if the filter doesn't find anything, gcloud exits
gracefully and the script can continue the loop waiting for the VM to
exist.

* Uses proper gcloud filter and removes --zone flag

* Fixes gcloud filter expression

* Adds cert_key to project metadata

Previously, the cert_key was added directly to the "cluster_data"
metadata entry for non-init control plane machines. However, this
required the init control plane machine script to loop, waiting for
those other machines to be created. Additionally, the process of adding
the cert_key to an existing metadata field was rather ungainly and ugly.
I consider adding the cert_key to project metadata to be a relatively
safe operation. See my comment in create-control-plane.sh for rationale.

* Fixes a typo and grammar error in comments

* Read/write kubeadm cert_key to/from SecretManager

Instead of writing a sensitive key to the project metadata, the "init"
control plane machine will write it to a secret in SecretManager. "join"
machines will read the secret from SecretManager.