Adds new dhclient exit hook script
The script should undo what Google's google_set_hostname exit hook does,
which is to set the hostname to anything up to the first dot, which we
don't want. We always want the hostname to be the FQDN, since this is
what kubeadm uses for the node name if not specified by the --node-name
flag.