Set hostname to k8s node name (#244)
* Removes --node-name flag from kubeadm for VMs

* Adds new dhclient exit hook script

The script should undo what Google's google_set_hostname exit hook does,
which is to set the hostname to anything up to the first dot, which we
don't want. We always want the hostname to be the FQDN, since this is
what kubeadm uses for the node name if not specified by the --node-name
flag.

* Removes dhclient exit hook script

It didn't work for some reason.

* Sets system's hostname in join-cluster.sh script

This should allow kubeadm to automatically pick the right node name,
since by default kubeadm uses the system's hostname as the node name.