Update extension-server, change URL
This commit updates the extension-server container image to v0.5.0,
which implements the new functionality to allow machines to request to
be removed from the cluster when they are rebooting or being shutdown.

This new extension has a new URL of /v1/node/delete, and this commit
also updates the leave-cluster.sh script to use this new path.