Populates new instance-type metadata file
We want an explicit way to know whether tests are originating from a
standard GCE VM, or one that is part of a MIG. This change updates the
write-metadatda.sh script to include a check for whether this instances
is part of a MIG or not. If it is, it writes "mig" to the file
/var/local/metadata/instance-type, else it writes "standard" to the
file.