Add new "instance-type" metadata file to virtual machines (#246)
* Populates new instance-type metadata file

We want an explicit way to know whether tests are originating from a
standard GCE VM, or one that is part of a MIG. This change updates the
write-metadatda.sh script to include a check for whether this instances
is part of a MIG or not. If it is, it writes "mig" to the file
/var/local/metadata/instance-type, else it writes "standard" to the
file.

* Fixes metadata URL

* Changes instance-type to loadbalanced and standalone

* Changes instance-type to loadbalanced

This change will create a new metadata file named "loadbalanced" with
content of "true" in the case that this is a MIG instance which is
loadbalanced.