Uses create-by instance attribute to identify MIG
Google says you can look for the presence of attributes
"instance-template" and "created-by" to determine if a VM is part of a
MIG. This commit changes how write-metadata.sh determines if the
instance is part of a MIG by checking "created-by" instead of
"instance-template". "created-by" specifically references the instance
group manager that created the instance, whereas "instance-template"
just lists the template that was used to create the instance. We may at
some point (should?) use instance templates to create regular standalone
VM, in which case GCP _could_ populate the "instance-template" attribute
even for non-MIG instances. This small change seems more future proof.

https://cloud.google.com/compute/docs/instance-groups/getting-info-about-migs#checking_if_a_vm_instance_is_part_of_a_mig