Adds NoPivotRoot=true to containerd config
This setting is necessary on our physical machines because they run from an
initramfs. Also, because this isn't wanted on virtual nodes, the commit also
restores the full config for VMs (no longer a symlink to the physical one), and
I now recall why there were two separate files to begin with.