Fixes env variables + updates containerd configs (#250)
* Removes writing ENV vars from control plane script

The packer build was already adding these vars to .bashrc, and the control
plane script was duplicating them. Additionally, write the same configs to
/root/.profile so that cron jobs can have access to the vars.

* Fixes some duplication with ENV variables

Multiple variables were being set twice.

* Updates containerd config

Updates the containerd config for stage3_ubuntu images, and symlinks to this
same file from the virtual_ubuntu images, since we always want them to be the
same.

* Adds NoPivotRoot=true to containerd config

This setting is necessary on our physical machines because they run from an
initramfs. Also, because this isn't wanted on virtual nodes, the commit also
restores the full config for VMs (no longer a symlink to the physical one), and
I now recall why there were two separate files to begin with.

* Moves containerd root and state to /cache

I forgot to port this config from the old one. Without it, containerd will
write all of its data to the root directory and fill up the filesystem.

* Allow CIDRs other than /26 in mlab-sandbox

We need to be able to test with small prefixes, and this change allows any
prefix but only in mlab-sandbox

* Adds clarifying comments to a few files.