Configure the join-cluster.service systemd unit to always restart  (#254)
* Always restart the join-cluster unit

If the join-cluster script fails, always restart it. It may continue to fail,
but other conditions that are causing it to fail may resolve themselves, in
which case having it always restart will allow the join operation to succeed
eventually.

* Configures kubeadm to ignore all preflight errors

* Increases RestartSec from 10 to 30 for join-cluster.service

In case a bunch of machines are restarting around the same time for some
unlikely reason, this might possibly help to lessen the load on the API
cluster.