Writes LB IPs to metadata directory if VM is a MIG instance (#255)
Depending on whether the VM is standalone or part of a MIG, we need to write
different values to /var/local/metadata/external_ip[v6]. For standalone VMs we
can write the same values that we have been writing before, which are the
external IPs associated with the machine. For MIG instances, we want to write
the external IPs associated with the loadbalancer instead.