Adds a new "managed" metadata file
The content of this file will indicate which parts of the infrastructure are
managed by M-Lab. In the case of a "full" site, M-Lab manages both the switch
and machine. For "minimal" deployments M-Lab manages only the machine. For BYOS
deployments, M-Lab manages nothing. For virtual machines M-Lab manages only the
machine.