Writes out new "managed" metadata file with appropriate value (#257)
* Adds a new "managed" metadata file

The content of this file will indicate which parts of the infrastructure are
managed by M-Lab. In the case of a "full" site, M-Lab manages both the switch
and machine. For "minimal" deployments M-Lab manages only the machine. For BYOS
deployments, M-Lab manages nothing. For virtual machines M-Lab manages only the
machine.

* Runs write-metadata.server After network.target

write-metadata.service needs to query information from device eth0, so
networking must be up and configured for this to work.

* Adds After=systemd-networkd-wait-online.service

This is an attempt to be sure that eth0 exists before the
write-metadata.service unit runs.

* Makes sure eth0 device exists before running write-metadata.service

* Calculates prefix length in a more robust way

No matter how I tried to order the write-metadata.service systemd unit, eth0
was always not present when the service ran. This change extracts the prefix
length from /proc/cmdline.

* Removes comment that is no longer necessary

* Removes the "none" option for the metadata file "managed"

The "none" option was supposed to account for BYOS deployments, but this repo
and script will not be used for those deployments.