Converts comma in managed value to a dash
For "full" physical sites the content of /var/local/metadata/managed is
"switch,machine". The plan was to use the value of this file as the value of a
new "mlab/managed" node label. However, kubernetes node label values cannot
contain commas. As a workaround, this commit changes any commas to dashes,
which are valid.