Causes Packer to use VMs internal IP address
We lately disabled SSH access to project VMs in most VPCs, notwithstanding the
mlab-platform-network. We don't SSH into those nodes, and they don't need to be
publicly exposed in that way. However, doing this broke epoxy-images builds for
Packer, because, by default, Packer will SSH into the VM that is created using
the public address. This commit modifies Packer to omit giving the VM a public
address, and to use the private internal address when SSHing into the VM to
provision the image. This should work fine for us, since Packer is running in
Cloud Build on the same VPC as the VM that is created.