Add network tags to VMs created by Packer for firewalling (#263)
* Causes Packer to use VMs internal IP address

We lately disabled SSH access to project VMs in most VPCs, notwithstanding the
mlab-platform-network. We don't SSH into those nodes, and they don't need to be
publicly exposed in that way. However, doing this broke epoxy-images builds for
Packer, because, by default, Packer will SSH into the VM that is created using
the public address. This commit modifies Packer to omit giving the VM a public
address, and to use the private internal address when SSHing into the VM to
provision the image. This should work fine for us, since Packer is running in
Cloud Build on the same VPC as the VM that is created.

* Adds network tag to Packer VM for firewalling

Removes the use_internal_ip and omit_public_ip options, since I did not
realize before that Cloud Build instances run in a Google-managed pool of
machines outside of project VPCs, so SSHing to the private IP of a VM created
by Packer will not work.

Instead, this new commit adds a network tag to the VM, which should allow me to
create a targeted firewall rule explicitly allowing SSH access to VMs created
by Packer.