Configure Packer to ssh to VMs using IAP (#264)
* Set googlecompute builder setting use_iap=true

Previously I had set up a special "cloud-build-packer" network tag on the VM
that packer created, and then created a special firewall rule to allow SSH to
VMs with that tag. This was overly specific to Packer, and wasn't general
enough for other use cases where we need/want SSH access into VMs. Instead, we
now set the use_iap=true setting, which will cause packer to tunnel through
Google IAP (Identity Aware Proxy), which is used for this very purpose. Now,
there will be a generic firewall rule allow SSH access from IAP IPs.

* Updates Packer build config to use Artifact Registry and v1.2

Container Registry is deprecated and going away. I'm taking this opportunity to
update the build image URL to that of Artifact Registry. This also updates the
image version to 1.2 which includes some changes to the image that Packer
needs.

* Sets iap_tunnel_launch_wait=60s

IAP is currently not working. Testing this to see if it perhaps resolves the
issue.

* Adds debug logging to Packer, temporarily, to debug ssh via IAP

* Forces Packer to use instance internal IP address

Maybe this will make ssh access via IAP work?

* Adds PACKER_LOG=1 to enable debug logging of Packer

* Adds Google Cloud SDK install dir to Packer's PATH

* Disables debug, propagate normal instances configs to others

I think this is now working. Largely the underlying issue was that Packer was
unable to find gcloud. Adding the Google SDK path to the environments PATH
fixed this. It turns out that using IAP also allows us to create the VM with no
public IP, which seems better.

* Removes omit_external_ip.. needed to install packages