Adds loadbalanced=false to physical machines and standalone VMs (#266)
We have been adding the metadata file "loadbalanced" with a content/value of
"true" to MIG instances. However, we have not been explicitly flagging
non-loadbalanced machines as loadbalanced=false. This commit makes the
declaration explicity for all machine types, which should make filtering in
BigQuery using this field more intuitive and consitent with other fields.