Adds CNI config files to root fs of all machine images
Previously, we had a weird configuration in the k8s platform cluster with two
separate flannel DaemonSets, one for physical machines and one for virtual
machines. The whole point of that was writing out different CNI config files
depending on the machine type. This commit just bakes those CNI configs into
the filesystem to avoid that extra complexity in the cluster. Those files
haven't changed in years.