Makes configure-tc-fq.sh agnostic to the speed of the uplink.
Having this script enforce possible uplink speed values makes it brittle. It
also introduces a second place where uplink speeds must be configured, in
addition to siteinfo, which is not ideal. Instead, this script shouldn't care
what the uplink value is, as long as it looks like a valid speed specification.