Make configure-tc-fq.sh agnostic to the speed of the uplink (#271)
* Adds 40g as a possible uplink speed in configure-tc-fq.sh

The new Ziply sites will be 40Gbps, and the script that configures the "fq"
qdisc on boot needs to know about this speed.

* Makes configure-tc-fq.sh agnostic to the speed of the uplink.

Having this script enforce possible uplink speed values makes it brittle. It
also introduces a second place where uplink speeds must be configured, in
addition to siteinfo, which is not ideal. Instead, this script shouldn't care
what the uplink value is, as long as it looks like a valid speed specification.