Dynamically determines eth* device with a layer 2 link
Previously, our network config hardcoded eth0 as the network device to
configure. This has worked until today, but is failing with our new Ziply
sites. For the Mellanox cards in the Ziply equipment, the kernel is not
consistently assigning eth* device names to the 6 interfaces on the NIC,
causing network breaks randomly between boots.

This commit introduces a small for loop that looks over all the eth*
interfaces, looking for the one with a layer 2 link, and uses that for the
network configuration.